Start rendering custom diff hunk controls for AssistantDiff
